# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ISKCON Purposes API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Purpose {
  id: number;
  number: number;
  title: string;
  shortDescription: string;
  fullDescription: string;
  icon: string;
}

export interface Activity {
  id: number;
  purposeId: number;
  title: string;
  description: string;
  authorName: string;
  createdAt: string;
}

export interface CreateActivityInput {
  title: string;
  description: string;
  authorName: string;
}

export interface Message {
  id: number;
  purposeId: number;
  content: string;
  authorName: string;
  createdAt: string;
}

export interface CreateMessageInput {
  content: string;
  authorName: string;
}