# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ISKCON Purposes API
 * OpenAPI spec version: 0.1.0
 */

export interface Activity {
  id: number;
  purposeId: number;
  title: string;
  description: string;
  authorName: string;
  createdAt: Date;
}