# lib/api-zod/src/generated/types/createActivityInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ISKCON Purposes API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateActivityInput {
  title: string;
  description: string;
  authorName: string;
}