# lib/api-zod/src/generated/types/createMessageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ISKCON Purposes API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMessageInput {
  content: string;
  authorName: string;
}