# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ISKCON Purposes API
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  purposeId: number;
  content: string;
  authorName: string;
  createdAt: Date;
}