# lib/api-zod/src/generated/types/purpose.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ISKCON Purposes API
 * OpenAPI spec version: 0.1.0
 */

export interface Purpose {
  id: number;
  number: number;
  title: string;
  shortDescription: string;
  fullDescription: string;
  icon: string;
}